+++ js/firesight/CalcOffset.js
@@ -14,12 +14,13 @@ var should = require("should");
 
         return that;
     }
-    CalcOffset.prototype.calculate = function(camName, onSuccess, onFail) {
+    CalcOffset.prototype.calculate = function(camName, onSuccess, onFail, options) {
         var that = this;
+        options = options || {};
         var firesight = that.firesight;
         camName = typeof camName == "undefined" ? firesight.camera.name : camName;
         var loc = firesight.images.location();
-        var savedImage = firesight.images.savedImagePath(camName);
+        var savedImage = firesight.images.savedImagePath(camName, options.savedImage);
         if (!savedImage) {
             onFail(new Error("FireSightREST.calcOffset() no saved image"));
             return that;